# ai-video-studio/codex/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Studio API",
    "version": "0.1.0",
    "description": "Semantic API contract for the Cutpilot/AI Video Studio MVP. Provider/model names are internal only."
  },
  "servers": [
    { "url": "https://api.example.local" }
  ],
  "paths": {
...
    "/projects/{projectId}/image-jobs": {
      "post": {
        "operationId": "createImageJob",
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^prj_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt", "purpose", "role", "aspect"],
                "additionalProperties": false,
                "properties": {
                  "prompt": { "type": "string" },
                  "purpose": { "$ref": "../schemas/domain.schema.json#/$defs/ImageMakerPurpose" },
                  "role": { "$ref": "../schemas/domain.schema.json#/$defs/ImageAssetRole" },
                  "aspect": { "$ref": "../schemas/domain.schema.json#/$defs/Aspect" },
                  "style": { "type": "string" },
                  "count": { "type": "integer", "minimum": 1, "maximum": 4 }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Image generation job queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["job"],
                  "additionalProperties": false,
                  "properties": {
                    "job": { "$ref": "../schemas/domain.schema.json#/$defs/ImageJob" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid image generation request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient credits to reserve this work",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/InsufficientCreditsResponse" }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/shots/{shotId}/generate": {
      "post": {
        "operationId": "generateShot",
        "parameters": [
          { "name": "shotId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^sht_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "tier": { "$ref": "../schemas/domain.schema.json#/$defs/Tier" },
                  "takeCount": { "type": "integer", "minimum": 1, "maximum": 3 }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Generation jobs queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "takes": { "type": "array", "items": { "$ref": "../schemas/domain.schema.json#/$defs/Take" } },
                    "jobs": { "type": "array", "items": { "$ref": "../schemas/domain.schema.json#/$defs/GenerationJob" } }
                  },
                  "required": ["takes", "jobs"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid generation request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient credits to reserve this work",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/InsufficientCreditsResponse" }
              }
            }
          },
          "404": {
            "description": "Shot not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/shots/{shotId}/references": {
      "post": {
        "operationId": "attachImageToShot",
        "parameters": [
          { "name": "shotId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^sht_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assetId", "mode"],
                "additionalProperties": false,
                "properties": {
                  "assetId": { "type": "string", "pattern": "^img_" },
                  "mode": { "$ref": "../schemas/domain.schema.json#/$defs/AssetUsageMode" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reference image attached to shot",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/Shot" }
              }
            }
          },
          "400": {
            "description": "Invalid shot reference image request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Shot or image asset not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/generate-all": {
      "post": {
        "operationId": "generateAll",
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^prj_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "tier": { "$ref": "../schemas/domain.schema.json#/$defs/Tier" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Project generation queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "jobs": { "type": "array", "items": { "$ref": "../schemas/domain.schema.json#/$defs/GenerationJob" } }
                  },
                  "required": ["jobs"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid project generation request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient credits to reserve this work",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/InsufficientCreditsResponse" }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/renders": {
      "post": {
        "operationId": "startRender",
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^prj_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["specs"],
                "additionalProperties": false,
                "properties": {
                  "specs": {
                    "type": "array",
                    "items": { "$ref": "../schemas/domain.schema.json#/$defs/ExportSpec" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Render jobs queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": { "$ref": "../schemas/domain.schema.json#/$defs/RenderJob" }
                    }
                  },
                  "required": ["jobs"]
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid render specs",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient credits to reserve this work",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/InsufficientCreditsResponse" }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Equivalent render work is already active",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}